Added specifications for Mesh and Skeletal data, as well as updated Mesh Bone Weight Attribute's size to reflect an UINT_16 rather than a UINT_32 for the bone index

+++ src/crucible/core/Mesh.cpp
@@ -197,7 +197,7 @@ namespace crucible
                 case COLOR:
                     return sizeof(char)*4;
                 case BONE_WEIGHT:
-                    return (sizeof(float)*4)+(sizeof(uint32_t)*4);
+                    return (sizeof(float)*4)+(sizeof(uint16_t)*4);
                 case VERTEX_ATTRIBUTE_ENUM_SIZE:
                     throw std::runtime_error("Invalid attribute provided");
             }
